chore: move out the rest of ryot to sub crates
