+++ crates/ryot/Cargo.toml
@@ -13,72 +13,22 @@ categories = ["game-development"]
 readme = "../../README.md"
 
 [dependencies]
-async-std = "1.12.0"
-derive_more = "0.99.17"
-serde_repr = "0.1"
-ryot_assets = { path = "../ryot_assets", version = "0.2" }
-ryot_content = { path = "../ryot_content", version = "0.2" }
-ryot_core = { path = "../ryot_core", version = "0.2" }
-ryot_tiled = { path = "../ryot_tiled", version = "0.2" }
-ryot_sprites = { path = "../ryot_sprites", version = "0.2" }
 ryot_internal = { path = "../ryot_internal", version = "0.2" }
-
-bevy = { workspace = true, optional = true }
-bevy_asset_loader = { workspace = true, optional = true }
-bevy_common_assets = { workspace = true, optional = true }
-bevy_stroked_text = { workspace = true, optional = true }
-bytes.workspace = true
-color-eyre.workspace = true
-egui = { workspace = true, optional = true }
-glam.workspace = true
-heed = { workspace = true, optional = true }
-image.workspace = true
-itertools.workspace = true
-leafwing-input-manager = { workspace = true, optional = true }
-log.workspace = true
-lzma-rs.workspace = true
-postcard = { workspace = true, optional = true }
-prost-types.workspace = true
-prost.workspace = true
-rand.workspace = true
-rayon.workspace = true
-serde.workspace = true
-serde_json.workspace = true
-strum.workspace = true
-thiserror.workspace = true
-time-test.workspace = true
-wasm-bindgen-futures.workspace = true
-zstd = { workspace = true, optional = true }
-
-[dev-dependencies]
-rstest.workspace = true
-quickcheck.workspace = true
-quickcheck_macros.workspace = true
 
 [features]
 default = [
-    "tibia",
     "bevy",
+    "tibia",
 ]
 
-bevy = [
-    "dep:bevy",
-    "dep:bevy_common_assets",
-    "dep:bevy_asset_loader",
-    "dep:bevy_stroked_text",
-    "dep:leafwing-input-manager",
-    "ryot_internal/bevy"
-]
+bevy = ["ryot_internal/bevy"]
 
-tibia = [
-    "ryot_assets/ryot_tibia",
-    "ryot_internal/ryot_tibia",
-]
+tibia = ["ryot_internal/ryot_tibia"]
 
-debug = ["dep:bevy_stroked_text", "ryot_internal/debug"]
-lmdb = ["dep:heed", "dep:postcard", "ryot_internal/lmdb"]
+debug = ["ryot_internal/debug"]
+lmdb = ["ryot_internal/lmdb"]
 compression = ["ryot_internal/compression"]
-egui = ["dep:egui", "ryot_internal/egui"]
+egui = ["ryot_internal/egui"]
 
 pathfinder = ["ryot_internal/ryot_pathfinder"]
 tiled_pathfinder = ["ryot_internal/tiled", "ryot_internal/ryot_pathfinder"]
