--- crates/ryot/src/bevy_ryot/game/mod.rs
@@ -1,1 +1,0 @@
-pub mod tile_flags;
