--- crates/ryot/src/bevy_ryot/mod.rs
@@ -1,2 +1,0 @@
-mod game;
-pub use game::*;
