+++ crates/ryot_app/src/game.rs
@@ -1,5 +1,6 @@
-use bevy_app::{App, Last, Plugin, Update};
+use bevy_app::{App, Last, Plugin, PostUpdate, Update};
 use bevy_ecs::prelude::*;
+use ryot_core::cache::Cache;
 
 use ryot_sprites::SpriteSystems;
 use ryot_tiled::prelude::*;
@@ -24,5 +25,18 @@ impl Plugin for ElevationPlugin {
             ),
         )
         .add_systems(Last, track_position_changes);
+    }
+}
+
+/// `TileFlagPlugin` provides the necessary system and resource setup for managing `TileFlags`
+/// within the game world. It ensures that the flag cache is up-to-date and reflects the latest
+/// flag state of the whole tile, per position. This avoids the need to iterate over each entity
+/// within a tile to check its properties.
+pub struct TileFlagPlugin;
+
+impl Plugin for TileFlagPlugin {
+    fn build(&self, app: &mut App) {
+        app.init_resource::<Cache<TilePosition, TileFlags>>()
+            .add_systems(PostUpdate, update_tile_flag_cache);
     }
 }
