+++ crates/ryot_pathfinder/Cargo.toml
@@ -24,7 +24,7 @@ version = "4.9.1"
 [dependencies]
 bevy_app.workspace = true
 bevy_ecs.workspace = true
-bevy_tasks.workspace = true
+bevy_tasks = { workspace = true, features = ["multi-threaded"] }
 
 ryot_core = { path = "../ryot_core", version = "0.2" }
 ryot_tiled = { path = "../ryot_tiled", version = "0.2.0", optional = true }
